always build ld64, regardless if we are on a 64 bit system or not fixed an issue in include/foreign/i386/_types.h which prevented compiling on 32 bit systems

+++ cctools/include/foreign/i386/_types.h
@@ -43,8 +43,8 @@ typedef	short			__int16_t;
 typedef	unsigned short		__uint16_t;
 typedef int			__int32_t;
 typedef unsigned int		__uint32_t;
-typedef long int		__int64_t;
-typedef unsigned long int	__uint64_t;
+typedef long long int		__int64_t;
+typedef unsigned long long  int	__uint64_t;
 
 typedef long			__darwin_intptr_t;
 typedef unsigned int		__darwin_natural_t;
